+++ src/euroeval/benchmarker.py
@@ -85,9 +85,6 @@ class Benchmarker:
         run_with_cli: bool = False,
         requires_safetensors: bool = False,
         download_only: bool = False,
-        model_language: str | c.Sequence[str] | None = None,
-        dataset_language: str | c.Sequence[str] | None = None,
-        batch_size: int | None = None,
         max_context_length: int | None = None,
         vocabulary_size: int | None = None,
     ) -> None:
@@ -173,12 +170,6 @@ class Benchmarker:
             download_only:
                 Whether to only download models and datasets without performing any
                 benchmarking. Defaults to False.
-            model_language:
-                Deprecated argument. Please use `language` instead.
-            dataset_language:
-                Deprecated argument. Please use `language` instead.
-            batch_size:
-                Deprecated argument. Please use `finetuning_batch_size` instead.
             max_context_length:
                 Override for the maximum context length of the model. If None, the value
                 will be inferred automatically from the model. Defaults to None.
@@ -201,53 +192,6 @@ class Benchmarker:
             else:
                 msg += "the argument `download_only` was set to True."
             raise ValueError(msg)
-
-        # Deprecation warnings
-        if batch_size is not None:
-            if run_with_cli:
-                msg = (
-                    "The --batch-size option is deprecated and will be removed in a "
-                    "future version. Please use --finetuning-batch-size instead. "
-                    "Overwriting --finetuning-batch-size with the value from "
-                    "--batch-size."
-                )
-            else:
-                msg = (
-                    "The `batch_size` argument is deprecated and will be removed in a "
-                    "future version. Please use `finetuning_batch_size` instead. "
-                    "Overwriting `finetuning_batch_size` with the value from "
-                    "`batch_size`."
-                )
-            log(msg, level=logging.WARNING)
-            finetuning_batch_size = batch_size
-        if model_language is not None:
-            if run_with_cli:
-                msg = (
-                    "The --model-language option is deprecated and will be removed in "
-                    "a future version. Please use --language instead. Ignoring the "
-                    "--model-language value."
-                )
-            else:
-                msg = (
-                    "The `model_language` argument is deprecated and will be removed "
-                    "in a future version. Please use `language` instead. Ignoring the "
-                    "`model_language` value."
-                )
-            log(msg, level=logging.WARNING)
-        if dataset_language is not None:
-            if run_with_cli:
-                msg = (
-                    "The --dataset-language option is deprecated and will be removed "
-                    "in a future version. Please use --language instead. Ignoring the "
-                    "--dataset-language value."
-                )
-            else:
-                msg = (
-                    "The `dataset_language` argument is deprecated and will be removed "
-                    "in a future version. Please use `language` instead. Ignoring the "
-                    "`dataset_language` value."
-                )
-            log(msg, level=logging.WARNING)
 
         # If FULL_LOG has been set, then force verbose mode
         if os.getenv("FULL_LOG", "0") == "1":
@@ -410,9 +354,6 @@ class Benchmarker:
         force: bool | None = None,
         verbose: bool | None = None,
         debug: bool | None = None,
-        model_language: str | c.Sequence[str] | None = None,
-        dataset_language: str | c.Sequence[str] | None = None,
-        batch_size: int | None = None,
         max_context_length: int | None = None,
         vocabulary_size: int | None = None,
     ) -> c.Sequence[BenchmarkResult]:
@@ -517,12 +458,6 @@ class Benchmarker:
             debug:
                 Whether to output debug information. Defaults to the value specified
                 when initialising the benchmarker.
-            model_language:
-                Deprecated argument. Please use `language` instead.
-            dataset_language:
-                Deprecated argument. Please use `language` instead.
-            batch_size:
-                Deprecated argument. Please use `finetuning_batch_size` instead.
             max_context_length:
                 Override for the maximum context length of the model. If None, the
                 value will be inferred automatically from the model. Defaults to the
@@ -549,31 +484,6 @@ class Benchmarker:
 
         if task is not None and dataset is not None:
             raise ValueError("Only one of `task` and `dataset` can be specified.")
-
-        # Deprecation warnings
-        if batch_size is not None:
-            log(
-                "The `batch_size` argument is deprecated and will be removed in a "
-                "future version. Please use `finetuning_batch_size` instead. "
-                "Overwriting `finetuning_batch_size` with the value from "
-                "`batch_size`.",
-                level=logging.WARNING,
-            )
-            finetuning_batch_size = batch_size
-        if model_language is not None:
-            log(
-                "The `model_language` argument is deprecated and will be removed "
-                "in a future version. Please use `language` instead. Ignoring the "
-                "`model_language` value.",
-                level=logging.WARNING,
-            )
-        if dataset_language is not None:
-            log(
-                "The `dataset_language` argument is deprecated and will be removed "
-                "in a future version. Please use `language` instead. Ignoring the "
-                "`dataset_language` value.",
-                level=logging.WARNING,
-            )
 
         # Get a new updated benchmark configuration, based on any changes to the
         # parameters
@@ -1185,25 +1095,6 @@ class Benchmarker:
                 f"Failed to benchmark model {model_config.model_id!r} on dataset "
                 f"{dataset_config.name!r} after {num_attempts} attempts."
             )
-
-    def __call__(self, *args: t.Any, **kwds: t.Any) -> t.Any:  # noqa: ANN401
-        """Alias for `self.benchmark()`.
-
-        Args:
-            *args:
-                Positional arguments to pass to `self.benchmark()`.
-            **kwds:
-                Keyword arguments to pass to `self.benchmark()`.
-
-        Returns:
-            The result of `self.benchmark()`.
-        """
-        log(
-            "Calling the `Benchmarker` class directly is deprecated. Please use the "
-            "`benchmark` function instead. This will be removed in a future version.",
-            level=logging.WARNING,
-        )
-        return self.benchmark(*args, **kwds)
 
 
 def get_record(